requirements.txt
robotframework
robotframework-seleniumlibrary
google-generativeai
beautifulsoup4
requests
gitpython
python-dotenv
